refactor(logging): move test subscriber + delete logging/telemetry from xmtp_common
Moves the test subscriber to xmtp_logging (removing the tracing-forest tree
logger + TestLogReplace), deletes the now-duplicated logging/telemetry modules
from xmtp_common, and drops the opentelemetry / tracing-subscriber production
deps + the 'logging' feature. tracing-subscriber survives only as a test/bench
dep (traced_test + bench subscriber). Consumers repointed to xmtp_logging in
follow-up changes.
