+++ crates/xmtp_logging/Cargo.toml
@@ -29,6 +29,8 @@ tracing-appender.workspace = true
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 tracing-web.workspace = true
+# test-utils only: readable panics in wasm test output.
+console_error_panic_hook = { workspace = true, optional = true }
 
 [target.'cfg(target_os = "android")'.dependencies]
 paranoid-android = "0.2"
@@ -38,4 +40,7 @@ tracing_android_trace = "0.1"
 tracing-oslog = "0.3"
 
 [features]
-test-utils = []
+# The test subscriber installed by `#[xmtp_common::test]`. On wasm this pulls in
+# `console_error_panic_hook` for readable panic output; on native it only needs
+# `tracing-subscriber`, which is already a non-optional dependency.
+test-utils = ["dep:console_error_panic_hook"]
